SetID conflicts resolve in expected ways now
SetID's method of removing the old listItem was broken sometimes due to the form refreshing immediately afterwards without the newly cached device information.
This is resolved by creating a new callback method that is passed when there is an ID conflict the user wants to override.
The callback method sets a member variable that gets read in the form's main loop that decides when to remove the selected device descriptor.
https://github.com/CrossTheRoadElectronics/Phoenix-IssueTracker/issues/567